In Geo::WebService::Elevation::USGS, only flush cached SOAP::Lite object in set() if the modified attribute is actually used when instantiating the SOAP::Lite object.
Add a localized $SOAP::Constants::DO_NOT_USE_CHARSET = 1; before the
actual SOAP call. This gets rid of the '; charset=utf-8' in the
Content-type header. I have had no trouble with this, but the SOAP::Lite
documentation recommends it, and it seems not only to have the desired
effect, but to work. I note, though, that the USGS is giving us this
same string back, even when we do not provide it.


git-svn-id: file:///Users/tom/Code/svn/repos/Geo-WebService-Elevation-USGS/trunk@1784 f6fd3633-1913-47c2-bad4-372d75c948ec